+++ include/utilization/utilization_session.h
@@ -9,14 +9,14 @@ namespace Utilization {
 	{
 		static const char *service_name() { return "utilization"; }
 
-		virtual double utilization(int core) = 0;
+		virtual int utilization(int core) = 0;
 
 
 		/*******************
 		 ** RPC interface **
 		 *******************/
 
-		GENODE_RPC(Rpc_utilization, double, utilization, int);
+		GENODE_RPC(Rpc_utilization, int, utilization, int);
 
 		GENODE_RPC_INTERFACE(Rpc_utilization);
 	};
